[db_tagwindow] Enable multiple selection in tag lists. (For drag and drop.)
This lets you copy many tags easily. But because GTK sucks, it might not
seem to work. You have to perform the drag as part of the same click that
selects the final tag you want, and you can't hold shift when you release
the mouse button (ctrl is ok though), or it doesn't work.

It's possible this is my fault, but I'm blaming it on GTK.