[fusefs] Fix (remove) threading.
I've been running it with -s, which works fine, but without it you can get
both a broken cache object (seems to happen in practice) and problems with
one thread reading a dbclient response meant for another (doesn't seem to
happen in practice).

If someone actually cares about threads they have to wrap the cache in a
lock (trivial), and either wrap dbclient in a lock (and lose pretty much
all the benefit of threads I think) or give each thread its own dbclient.
Since we don't want to make too many, that would actually mean keeping a
pool of dbclients (maybe 3) that threads can check out for use when they
need one. I can't imagine it's worth the effort.