[tagwindow] Ignore duplicate drags by looking at context.
The point was never to ignore tags that were there, just to avoid
duplicating all dropped tags, so hopefully this actually works (in
all GTK versions), without the problems of the old one. (Which as
implemented were that "bar" was ignored if "foobar" or "foo=bar" or
similar was the current tail, but this is kinda tricky to solve for
all cases with text handling.)

And I still have no idea why I get the events twice.