[add] avoid problems with identical files
you shouldn't of course, but if you do have several identical files and try
to add them at the same time, you could get errors (mostly from tagging) on
them because of races.

this can end up with a processes temporarily stuck getting an in-progress
post and putting it back in the queue, only to immediately get it again.
(but it stops once the other process processing an identical file finishes,
so it's not really a problem.)