[client] Add context manager support.
So you can do

	with Client() as client:
		...

and have it close automatically at the end. (If opened. Creating or
entering the client does not connect to the server, that only happens
when needed.)

Just

	with client:
		...

works too, and will still close the server connection at the end, but
using the client again will re-open it.